+++ packages/pose-core/src/types.ts
@@ -55,6 +55,7 @@ export interface Poser<V, A, C, P> {
   get: (key?: string) => any;
   has: (key: string) => boolean;
   setProps: (props: Props) => void;
+  setTransitionProps: (props: Props) => void;
   destroy: () => void;
   _addChild: (
     config: PoserConfig<V>,